+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Sync to upstream Luau 0.668
 
+### Fixed
+
+- Don't show function types with function kind or label details when autocompleting in a type
+  context ([#987](https://github.com/JohnnyMorganz/luau-lsp/issues/987))
+
 ## [1.42.0] - 2025-03-27
 
 ### Changed